requirements.txt
# Core testing framework
pytest>=7.1.2,<8.0.0
allure-pytest>=2.9.45,<3.0.0
allure-python-commons>=2.9.45,<3.0.0

# HTTP requests and web testing
requests>=2.26.0,<3.0.0
requests-toolbelt>=0.9.1,<1.0.0
urllib3>=1.26.7,<2.0.0

# Data processing and validation
PyYAML>=5.4.1,<7.0.0
pydantic>=1.8.2,<2.0.0
jsonpath>=0.82,<1.0.0

# Database connectivity
PyMySQL>=1.0.2,<2.0.0
redis>=4.3.4,<5.0.0

# Excel file processing
openpyxl>=3.0.9,<4.0.0
xlrd>=2.0.1,<3.0.0
xlutils>=2.0.0,<3.0.0
xlwt>=1.3.0,<2.0.0
xlwings>=0.27.7,<1.0.0

# Logging and output formatting
colorlog>=6.6.0,<7.0.0
colorama>=0.4.4,<1.0.0

# Notification services
DingtalkChatbot>=1.5.3,<2.0.0

# Cryptography and security
cryptography>=36.0.0,<40.0.0
pyOpenSSL>=21.0.0,<24.0.0
pyDes>=2.0.1,<3.0.0

# Network proxy and monitoring
mitmproxy>=8.1.0,<9.0.0

# Fake data generation
Faker>=9.8.3,<20.0.0

# Date and time utilities
python-dateutil>=2.8.2,<3.0.0

# Parallel testing
pytest-xdist>=2.4.0,<4.0.0
pytest-forked>=1.3.0,<2.0.0

# File and data processing
aiofiles>=0.8.0,<1.0.0
toml>=0.10.2,<1.0.0

# Web framework (if needed for mock services)
Flask>=2.0.3,<3.0.0
Jinja2>=3.1.2,<4.0.0

# Additional utilities
six>=1.16.0,<2.0.0
packaging>=21.3,<24.0.0
certifi>=2021.10.8

# Performance monitoring
psutil>=5.9.0,<6.0.0

# Code quality tools
flake8>=6.0.0,<7.0.0
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0
mypy>=1.0.0,<2.0.0
pre-commit>=3.0.0,<4.0.0

# Additional testing tools
pytest-html>=3.1.1,<4.0.0
pytest-cov>=4.0.0,<5.0.0
pytest-mock>=3.10.0,<4.0.0
pytest-timeout>=2.1.0,<3.0.0

# Enhanced logging
loguru>=0.6.0,<1.0.0

# Windows-specific dependencies (conditional)
pywin32>=304; sys_platform == "win32"
